AbaqusReader.jl documentation

AbaqusReader.jl documentation

AbaqusReader.jl documentation

Exported functions

Internal functions

Index

Exported functions

AbaqusReader.abaqus_download — Function.

abaqus_download(model_name; dryrun=false)

Download ABAQUS model from Internet. model_name is the name of the input file.

Given some model name from documentation, e.g., et22sfse, download that file to local file system. This function uses environment variables to determine the download url and place of storage.

In order to use this function, one must set environment variable ABAQUS_DOWNLOAD_URL, which determines a location where to download. For example, if the path to model is https://domain.com/v6.14/books/eif/et22sfse.inp, ABAQUS_DOWNLOAD_URL will be the basename of that path, i.e., https://domain.com/v6.14/books/eif.

By default, the model will be downloaded to current directory. If that is not desired, one can set another environment variable ABAQUS_DOWNLOAD_DIR, and in that case the file will be downloaded to that directory.

Function call will return full path to downloaded file or nothing, if download is failing because of missing environment variable ABAQUS_DOWNLOAD_DIR.

source

AbaqusReader.abaqus_read_mesh — Function.

abaqus_read_mesh(fn::String)

Read ABAQUS mesh from file fn. Returns a dict with elements, nodes, element sets, node sets and other topologically imporant things, but not the actual model with boundary conditions, load steps and so on.

source

AbaqusReader.abaqus_read_model — Function.

abaqus_read_model(filename::String)

Read ABAQUS model from file. Include also boundary conditions, load steps and so on. If only mesh is needed, it's better to use abaqus_read_mesh insted.

source

AbaqusReader.create_surface_elements — Function.

create_surface_elements(mesh, surface_name)

Create surface elements for surface using mesh mesh. Mesh can be obtained by using abaqus_read_mesh.

source

Internal functions

AbaqusReader.parse_definition — Method.

Parse string to get set type and name

source

AbaqusReader.parse_section — Function.

Parse nodes from the lines

source

Parse elements from input lines

Reads element ids and their connectivity nodes from input lines. If elset definition exists, also adds the set to model.

source

Parse node and elementset from input lines

source

Parse SURFACE keyword

source

AbaqusReader.regex_match — Function.

Custon regex to find match from string. Index used if there are multiple matches

source

AbaqusReader.add_set! — Function.

Add set to model, if set exists

source

AbaqusReader.consumeList — Function.

Custom list iterator

Simple iterator for comsuming element list. Depending on the used element, connectivity nodes might be listed in multiple lines, which is why iterator is used to handle this problem.

source

AbaqusReader.parse_numbers — Function.

Parse all the numbers from string

source

AbaqusReader.register_abaqus_keyword — Function.

register_abaqus_keyword(keyword::String)

Add ABAQUS keyword s to register. That is, after registration every time keyword show up in .inp file a new section is started

source

AbaqusReader.is_abaqus_keyword_registered — Function.

is_abaqus_keyword_registered(keyword::String)

Return true/false is ABAQUS keyword registered.

source

AbaqusReader.element_mapping — Constant.

element_mapping

This mapping table contains information what node ids locally match each side of element.

source

AbaqusReader.find_keywords — Function.

Find lines, which contain keywords, for example "*NODE"

source

AbaqusReader.matchset — Function.

Match words from both sides of '=' character

source

AbaqusReader.empty_or_comment_line — Function.

Checks for a comment or empty line

Function return true, if line starts with comment character "**" or has length of 0

source

AbaqusReader.create_surface_element — Function.

Given element code, element side and global connectivity, determine boundary element. E.g. for Tet4 we have 4 sides S1..S4 and boundary element is of type Tri3.

source

AbaqusReader.parse_abaqus — Function.

Main function for parsing Abaqus input file.

Function parses Abaqus input file and generates a dictionary of all the available keywords.

source

Index

AbaqusReader.element_mapping

AbaqusReader.abaqus_download

AbaqusReader.abaqus_read_mesh

AbaqusReader.abaqus_read_model

AbaqusReader.add_set!

AbaqusReader.consumeList

AbaqusReader.create_surface_element

AbaqusReader.create_surface_elements

AbaqusReader.empty_or_comment_line

AbaqusReader.find_keywords

AbaqusReader.is_abaqus_keyword_registered

AbaqusReader.matchset

AbaqusReader.parse_abaqus

AbaqusReader.parse_definition

AbaqusReader.parse_numbers

AbaqusReader.parse_section

AbaqusReader.regex_match

AbaqusReader.register_abaqus_keyword